requirements.txt
PyYAML==3.12
ruamel.yaml==0.15.39
youtube-dl==2018.6.14
